fighting!

+++ src/net/vizbits/gladiator/server/game/Gladiator.java
@@ -4,12 +4,12 @@ package net.vizbits.gladiator.server.game;
 public class Gladiator {
 
   public String name;
-  public Integer current_hp;
+  public Double current_hp;
   public Integer current_ap;
 
   public Integer character_level;
   public CharacterClass character_class;
-  public Integer max_hp;
+  public Double max_hp;
   public Integer max_ap;
   public Integer exp;
   public Integer attack;
